+++ packages/terra-dynamic-grid/src/terra-dev-site/doc/dynamic-grid/DynamicGrid.1.doc.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -24,6 +24,7 @@ import ResponsiveGridSrc from '!raw-loader!../../../../src/terra-dev-site/doc/ex
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[