+++ packages/terra-text/src/terra-dev-site/doc/text/Text.1.doc.jsx
@@ -8,7 +8,7 @@ import TextWeightDocs from '../../../../docs/text-weights.md';
 import TextVariationsDocs from '../../../../docs/text-variations.md';
 import TextVisuallyHiddenDocs from '../../../../docs/text-visually-hidden.md';
 import TextWordWrappedDocs from '../../../../docs/text-word-wrapped.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -34,6 +34,7 @@ import TextWordWrappedSrc from '!raw-loader!../../../../src/terra-dev-site/doc/e
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[