+++ packages/terra-signature/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Import for package version from package.json added to doc page
 
 2.18.0 - (August 14, 2019)
 ------------------